+++ _states/gpfs.py
@@ -74,3 +74,36 @@ def joined(cluster,
     ret['changes'] = changes
 
     return ret
+
+def started(runas=None,
+            **kwargs):
+
+    '''
+    Ensure the current node has started gpfs
+    '''
+
+    result = __salt__['gpfs.cluster_started'](runas=runas)
+    changes = {}
+    ret = { 'name' : __grains__['host'] }
+
+    if result == False:
+      if __opts__['test']:
+        ret['result'] = None
+        ret['comment'] = "GPFS would be started"
+      else:
+        res1 = __salt__['gpfs.start_cluster'](runas=runas)
+        if res1 == False:
+          ret['result'] = False
+          ret['comment'] = "There was an error start GPFS"
+        else:
+          changes['old'] = 'gpfs down'
+          changes['new'] = 'gpfs started'
+          ret['result'] = True
+          ret['comment'] = 'GPFS was started'
+    else:
+      ret['result'] = True
+      ret['comment'] = 'GPFS is already active'
+
+    ret['changes'] = changes
+
+    return ret