+++ CHANGELOG.rst
@@ -1,6 +1,11 @@
 gpfs formula
 ===========
 
+0.0.2 (TBC)
+
+- Add the capability of starting gpfs by using salt _modules and _states
+- Add capability of joining the cluster, if it's already not in the cluster using mmaddnode
+
 0.0.1 (2016-12-27)
 
 - Initial release taken from engineering/salt_configs